Added prevention of logging in with already logged in user's nick + chat history bugfix.
The first part forced some changes in which parts are done by server and which by client.
Now the client waits for server's signal if login was successful or nick got taken and
updates itself when it receives server side feedback.

As for history feature, it allowed scrolling into filler parts of history before they
actually got fileed with real user chat. That is now prevented and there won't be anything
displaying in history that the user themselves did not type.