OOP and event driven programming part I
I created first, rather crude Chat constructor and object.
For now the base role of it is to register listeners to chat events
and propagate events to them, so each listener handles it's behaviour itself
rather than having for example a click event on user list handle all the ui changes
in the whole application.

For now clicking on a user from list will trigger a privateChat event and have
chat object emit it to all listeners which then handle it.

Which atm is just the input object displaying name selected in console.